int64 indexing: SimdMatMul, np.nonzero, NDArray indexer, Transpose
- SimdMatMul.MatMulFloat accepts long M, N, K (validates <= int.MaxValue internally)
- MatMul2DKernel delegate uses long M, N, K
- np.nonzero returns NDArray<long>[] instead of NDArray<int>[]
- NDArray pointer indexer changed from int* to long*
- SwapAxes uses long[] for permutation

+++ src/NumSharp.Core/Backends/Default/ArrayManipulation/Default.Transpose.cs
@@ -82,7 +82,7 @@ namespace NumSharp.Backends
         public override NDArray SwapAxes(NDArray nd, int axis1, int axis2)
         {
             var ndims = nd.ndim;
-            var dims = new int[ndims];
+            var dims = new long[ndims];
             for (int i = 0; i < ndims; i++)
                 dims[i] = i;
